+++ src/body.cpp
@@ -11,5 +11,5 @@
 float DENSITY = 4E3;
 
 // Constructor definition
-Body::Body(const std::array<double, 3>* initial_position, const std::array<double, 3>* initial_velocity, double mass, float radius)
+Body::Body(const std::array<float, 3>* initial_position, const std::array<float, 3>* initial_velocity, float mass, float radius)
     : Particle(initial_position, initial_velocity, mass), radius((radius == 0) ? pow(3 * mass / (4 * M_PI * DENSITY), 1./3.) : radius) {};
